requirements.txt
mysql-connector-python~=9.3.0
python-dotenv~=1.1.1
fastapi~=0.116.1
starlette~=0.47.1
pydantic~=2.11.7
uvicorn[standard]~=0.34.3
